--- a/internal/codeintel/ranking/indexer_test.go
+++ b/internal/codeintel/ranking/indexer_test.go
@@ -14,8 +14,6 @@ import (
 )
 
 func TestIndexRepository(t *testing.T) {
-	t.Skip() // Flaky
-
 	ctx := context.Background()
 	mockStore := NewMockStore()
 	gitserverClient := NewMockGitserverClient()