Following up on https://github.com/sourcegraph/sourcegraph/pull/42931#pullrequestreview-1141880082 and https://github.com/sourcegraph/sourcegraph/pull/42931#discussion_r995387498.

This PR replaces some of the occurrence search machinery with Aho-Corasick for string search.

| repo                                 | old elapsed | new elapsed | delta           |
| ------------------------------------ | ----------- | ----------- | --------------- |
| github.com/hashicorp/errwrap         |  20.89ms    |  16.37ms    | 21.63% decrease |
| github.com/hashicorp/go-multierror   |  12.10ms    |  12.55ms    |  3.71% increase |
| github.com/sourcegraph/lsif-go       |  23.56ms    |  24.24ms    |  2.88% increase |
| github.com/sourcegraph-testing/etcd  | 106.92ms    | 105.79ms    |  1.05% decrease |
| github.com/sourcegraph-testing/tidb  |  89.54ms    | 100.78ms    | 12.55% increase |
| github.com/sourcegraph-testing/titan |  18.79ms    |  21.52ms    | 14.52% increase |
| github.com/sourcegraph-testing/zap   |  21.72ms    |  29.25ms    | 34.66% increase |
| github.com/sourcegraph/sourcegraph   | 149.12ms    | 218.26ms    | 46.36% increase |
| github.com/vishvananda/netlink       |  19.89ms    |  25.98ms    | 30.61% increase |

In the original approach, I split the _document_ into words over a linear pass, then compared those words against symbols. In this approach, I create a Aho-Corsaick trie out of the search symbols, then run it over the document. In both approaches we have a linear search over the document, but in the former we break all of the possible occurrences on word boundaries; on the later we only look for substrings. There is a difference in semantics here, and I'm sure we can account for that.

It's also slower, but we can probably optimize whatever is the issue here. I think the approach is probably sound if we can address the problem above.

## Test plan

N/A.