@keegancsmith 2022-10-17
The code looks good. Is it possible keeping the word boundary splitting + allEmitted code would help with perf? I am wondering if using the matcher just leads us to finding _lots_ of substrings that are already emitted which dominates perf.

I also had a shower thought to speed up the main branch. Can we replace the loop over searchTermAndPaths and just look at each symbol and do a hashtable lookup?