@efritz 2022-10-17
Let's not hyper optimize this part of the code, as well hopefully be using higher-quality data than diverting search where possible.

This is only a rough PoC to show in-repo file ranking works.